@@ -7,10 +7,12 @@ namespace App\Application\Http\Handler;
 use Antidot\React\PSR15\Response\PromiseResponse;
 use App\Domain\Model\TodoId;
 use App\Domain\TodoRepository;
+use Laminas\Diactoros\Response\RedirectResponse;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Zend\Diactoros\Response\RedirectResponse;
+
+use function React\Promise\resolve;
 
 class RemoveTodo implements RequestHandlerInterface
 {
@@ -29,12 +31,11 @@ class RemoveTodo implements RequestHandlerInterface
      */
     public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        return new PromiseResponse($this->repository->remove(TodoId::fromString(
-            $request->getAttribute('todo_id')
-        ))
-            ->then(static function () {
-                return new RedirectResponse('/');
-            })
-        );
+        return new PromiseResponse(resolve(
+            $this->repository->remove(TodoId::fromString(
+                $request->getAttribute('todo_id')
+            ))
+            ->then(static fn() => new RedirectResponse('/'))
+        ));
     }
 }
