@@ -1,8 +1,8 @@
 <?php
 
 // Load configuration
-use Antidot\Container\ContainerBuilder;
+use Antidot\Container\Builder;
 
 $config = require __DIR__ . '/../config/cli-config.php';
 
-return ContainerBuilder::build($config, true);
+return Builder::build($config, true);
